Add files via upload
READ ME! So far I can get the code to show the entire .txt file of inventory. I'm trying to get it to write to a new file with the line of the product ID and the 4 other fields.  I have not reached this functionality yet, just a start off point that I want to upload to the repository for more iteration